% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia_palette.R
\name{arcadia_palette}
\alias{arcadia_palette}
\title{Arcadia color palettes}
\usage{
arcadia_palette(palette_name)
}
\arguments{
\item{palette_name}{A character string specifying the name of the palette to retrieve.
Possible values are "Neutral", "Accent", "Lighter_accents", "Accent_expanded", "Lighter_accent_expanded".}
}
\value{
A function that can be used to generate Arcadia palettes with different options
}
\description{
Arcadia color palettes
}
\examples{
arcadia_palette("Neutral")
}
