% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plot.R
\name{clean_plot}
\alias{clean_plot}
\title{Clean Arcadia Plot}
\usage{
clean_plot()
}
\value{
A ggplot object with capitalized axis titles
}
\description{
Generalized extension function for ggplot cleanup. Right now, cleans up axis titles and removes axis lines and whitespace depending on geom type.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point() + labs(x = "horsepower", y = "miles per gallon") + clean_plot()
}
}
