% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fonts.R
\name{load_arcadia_fonts}
\alias{load_arcadia_fonts}
\title{Load in custom fonts}
\usage{
load_arcadia_fonts(custom_font = "Suisse", fallback_font = "sans")
}
\arguments{
\item{custom_font}{Font family name to be loaded in}

\item{fallback_font}{Font family name to be used in case custom font not available}
}
\value{
Custom font families or fallback font to use in plots.
}
\description{
Load in custom fonts
}
