% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_arcadia_plot.R
\name{save_arcadia_plot}
\alias{save_arcadia_plot}
\title{Save Arcadia Plot}
\usage{
save_arcadia_plot(filepath, plot, orientation = c("landscape", "portrait"))
}
\arguments{
\item{filepath}{The path where the plot will be saved}

\item{plot}{ggplot object to be saved}

\item{orientation}{The orientation of the plot, either "landscape" or "portrait"}
}
\value{
None
}
\description{
Save Arcadia Plot
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point()
save_arcadia_plot(p, "plot.pdf", "landscape")
}
}
