% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia_palette.R
\name{show_arcadia_gradients}
\alias{show_arcadia_gradients}
\title{Show gradient color palettes as a vector of hex codes}
\usage{
show_arcadia_gradients()
}
\value{
Gradient palette vectors with hex codes
}
\description{
Show gradient color palettes as a vector of hex codes
}
\examples{
show_arcadia_gradients()
}
