% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia_palette.R
\name{show_arcadia_palettes}
\alias{show_arcadia_palettes}
\title{Show palette options}
\usage{
show_arcadia_palettes()
}
\value{
Palette vectors with hex codes
}
\description{
Show palette options
}
\examples{
show_arcadia_palettes()
}
