% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia_palette.R
\name{arcadia_gradient_palette}
\alias{arcadia_gradient_palette}
\title{Arcadia gradient color palettes function}
\usage{
arcadia_gradient_palette(gradient_name, reverse = FALSE)
}
\arguments{
\item{gradient_name}{A character string specifying the name of the gradient palette to retrieve. Possible values are elements in GRADIENTS}

\item{reverse}{#' @description}
}
\value{
A function that can be used to generate the Arcadia gradient palettes with different palette options.
}
\description{
Arcadia gradient color palettes function
}
\examples{
arcadia_gradient_palette("wine")
arcadia_gradient_palette("wine", reverse = TRUE)
}
