% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_arcadia_plot.R
\name{save_arcadia_plot}
\alias{save_arcadia_plot}
\title{Save Arcadia Plot}
\usage{
save_arcadia_plot(
  filepath,
  plot,
  panel_size = "full_square",
  transparent_bg = TRUE,
  formats = c("pdf")
)
}
\arguments{
\item{filepath}{The path where the plot will be saved without a file extension}

\item{plot}{ggplot object to be saved}

\item{panel_size}{The size of the panel, one of "full_wide", "float_wide", "half_square", "full_square", "float_square". Default is "full_square"}

\item{transparent_bg}{Logical, whether to set the plot background to be transparent}

\item{formats}{Vector of file formats to export the plot as. Default is "pdf". Can be "pdf", "png", or both.}
}
\value{
None
}
\description{
Save Arcadia Plot
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point()
save_arcadia_plot("plot_full_wide", p, "full_wide")
save_arcadia_plot("/path/to/figs/folder/plot_float_wide", p, "float_wide", formats = c("pdf", "png"))
save_arcadia_plot("plot_half_square", p, "half_square", formats = "png")
save_arcadia_plot("plot_full_square", p, "full_square")
save_arcadia_plot("plot_float_square", p, "float_square")
}
}
