% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_arcadia_plot.R
\name{save_arcadia_plot}
\alias{save_arcadia_plot}
\title{Save Arcadia Plot}
\usage{
save_arcadia_plot(
  filepath,
  plot,
  panel_size = c("full_wide", "float_wide", "half_square", "full_square", "float_square")
)
}
\arguments{
\item{filepath}{The path where the plot will be saved}

\item{plot}{ggplot object to be saved}

\item{panel_size}{The size of the panel, one of "full_wide", "float_wide", "half_square", "full_square", "float_square"}

\item{transparent_bg}{Logical, whether to set the plot background to be transparent}
}
\value{
None
}
\description{
Save Arcadia Plot
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point()
save_arcadia_plot("plot_full_wide.pdf", p, "full_wide")
save_arcadia_plot("plot_float_wide.pdf", p, "float_wide")
save_arcadia_plot("plot_half_square.pdf", p, "half_square"
save_arcadia_plot("plot_full_square.pdf", p, "full_square")
save_arcadia_plot("plot_float_square.pdf", p, "float_square")
}
}
