% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia_palette.R
\name{show_arcadia_palettes}
\alias{show_arcadia_palettes}
\title{Show main palette options}
\usage{
show_arcadia_palettes()
}
\value{
Palette vectors with hex codes.
}
\description{
Shows the main color palette vectors with hex codes.
}
\examples{
show_arcadia_palettes()
}
