% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia.R
\name{theme_arcadia}
\alias{theme_arcadia}
\title{Arcadia theme adhering to illustration style guides}
\usage{
theme_arcadia(
  x_axis_type = "numerical",
  y_axis_type = "numerical",
  background = TRUE,
  padding = c(0.25, 0.25, 0.25, 0.25)
)
}
\arguments{
\item{x_axis_type}{Type of the x-axis data("categorical" or "numerical")}

\item{y_axis_type}{Type of the y-axis data("categorical" or "numerical")}

\item{background}{Default set to TRUE to have parchment background fill, set to FALSE to be transparent}

\item{padding}{Amount of padding white-space to add around the plot}

\item{base_size}{Default size of text in plot}
}
\value{
Layers onto an existing ggplot2 plot with a theme adhering to Arcadia style guidelines
}
\description{
Arcadia theme adhering to illustration style guides
}
\examples{

\dontrun{
ggplot(data=mtcars, aes(x=hp, y=mpg, color = as.factor(cyl))) +
  geom_point(size=2.5) +
  theme_arcadia()
}

\dontrun{
ggplot(data=mtcars, aes(x=hp, fill = as.factor(cyl))) +
  geom_density(alpha=0.8) +
  theme_arcadia()
}
}
