% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcadia.R
\name{theme_arcadia}
\alias{theme_arcadia}
\title{Arcadia theme adhering to illustration style guides}
\usage{
theme_arcadia(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Default size of test in plot}

\item{base_family}{Default font family in plot}
}
\value{
Reformatted ggplot with Arcadia Science style theme
}
\description{
Arcadia theme adhering to illustration style guides
}
